Multiple updates (#55)
* replace metadata file from txt to pkl

* waveform filename edits

* try implementing maya format checking on download

* add gitignore entry for .version

* add mayawaves to requirements

* edits to make things minimally workable

* add method to clear cache, rename pickle->pkl

* Update docstring for RIT catalog class

* Improve metadata handling for RIT

Some fields are often skipped in some simulations' metadata
Some fields are present but their values are left empty

We now insert numerical values in both of these scenarios,
the value we currently insert is zero (0). For some of the
fields, this value is the most reasonable one (e.g. for x,y
components of spin for non-spinning or aligned-spin configs,
and orbital eccentricity),
while for all others its a clearly absurd choice which would
indicate its artifical insertion

* Add docstring to Mayacatalog

* Prefer targeted import from sxs package

* Add doc to catalog.get and catalog.get_metadata methods

* add default values for total mass, and update docs

* Add infrastructure to download and cache RIT Psi4 data

* Add method to return a label for the modes being held

* Update script that downloads the entire RIT waveform catalog

* pre-commit fixes

---------

Co-authored-by: Aditya Vijaykumar <vijaykumar.aditya@gmail.com>